// types/GuitarsReponse.ts
// Generated by https://quicktype.io

export interface GuitarsResponse {
  data: GuitarsResponseDatum[]
  meta: Meta
}

export interface GuitarsResponseDatum {
  id: number
  attributes: PurpleAttributes
}

export interface PurpleAttributes {
  nombre: string
  descripcion: string
  precio: number
  url: string
  createdAt: string
  updatedAt: string
  publishedAt: string
  Imagen: Imagen
}

export interface Imagen {
  data: ImagenDatum[]
}

export interface ImagenDatum {
  id: number
  attributes: FluffyAttributes
}

export interface FluffyAttributes {
  name: string
  alternativeText: string
  caption: string
  width: number
  height: number
  formats: Formats
  hash: string
  ext: string
  mime: string
  size: number
  url: string
  previewUrl: null
  provider: string
  provider_metadata: ProviderMetadata
  createdAt: string
  updatedAt: string
}

export interface Formats {
  thumbnail: Large
  large: Large
  medium: Large
  small: Large
}

export interface Large {
  name: string
  hash: string
  ext: string
  mime: string
  path: null
  width: number
  height: number
  size: number
  url: string
  provider_metadata: ProviderMetadata
}

export interface ProviderMetadata {
  public_id: string
  resource_type: string
}

export interface Meta {
  pagination: Pagination
}

export interface Pagination {
  page: number
  pageSize: number
  pageCount: number
  total: number
}